Add files via upload
The Eliza program written in Prolog. It functions as a virtual therapist and can maintain a very basic conversation.

It works by reading in characters until it finds a space. The program recognized that as a "word". It then checks that word against a dictionary of responses and synonyms. If it finds a match, it prints out a defined response. It will continue along this path until it receives a keyword to make it stop.